#19 Get entitlements
